+++ tests/test_compressed_jwt.py
@@ -15,13 +15,13 @@ class CompressedPyJWT(PyJWT):
 def test_decodes_complete_valid_jwt_with_compressed_payload():
     # Test case from https://github.com/jpadilla/pyjwt/pull/753/files
     example_payload = {"hello": "world"}
-    example_secret = "secret"
+    example_secret = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
     # payload made with the pako (https://nodeca.github.io/pako/) library in Javascript:
     # Buffer.from(pako.deflateRaw('{"hello": "world"}')).toString('base64')
     example_jwt = (
-        b"eyJhbGciOiAiSFMyNTYiLCAidHlwIjogIkpXVCJ9"
-        b".q1bKSM3JyVeyUlAqzy/KSVGqBQA="
-        b".08wHYeuh1rJXmcBcMrz6NxmbxAnCQp2rGTKfRNIkxiw="
+        b"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
+        b".q1bKSM3JyVeyUlAqzy_KSVGqBQA"
+        b".vkKB9BEuLsUnHbA6GBhk2MlmBRZuzH8Fo2GmBqzFdgc"
     )
     decoded = CompressedPyJWT().decode_complete(
         example_jwt, example_secret, algorithms=["HS256"]
@@ -31,7 +31,7 @@ def test_decodes_complete_valid_jwt_with_compressed_payload():
         "header": {"alg": "HS256", "typ": "JWT"},
         "payload": example_payload,
         "signature": (
-            b"\xd3\xcc\x07a\xeb\xa1\xd6\xb2W\x99\xc0\\2\xbc\xfa7"
-            b"\x19\x9b\xc4\t\xc2B\x9d\xab\x192\x9fD\xd2$\xc6,"
+            b"\xbeB\x81\xf4\x11..\xc5'\x1d\xb0:\x18\x18d\xd8\xc9f\x05\x16n"
+            b"\xcc\x7f\x05\xa3a\xa6\x06\xac\xc5v\x07"
         ),
     }